etl/requirements.txt
pydantic==1.10.8
psycopg2-binary==2.9.10
pytz==2023.3
elasticsearch==7.17.12
redis==5.2.0